+++ python/graphstorm/inference/__init__.py
@@ -21,4 +21,3 @@ from .np_infer import GSgnnNodePredictionInferrer
 from .ep_infer import GSgnnEdgePredictionInferrer
 from .emb_infer import GSgnnEmbGenInferer
 from .mt_infer import GSgnnMultiTaskLearningInferrer
-from .utils import wrap_model_artifacts
